Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 #define PI 3.14
 int main(){
-    int r;
+    float r;
     float area;
-    scanf("%d",&r);
+    scanf("%f",&r);
     area=PI*r*r;
     printf("Area: %.2f",area);
     return 0;
